flutter refactor, most tests passing, UI issus
