+++ flutter_app/lib/features/ai_assistant/services/ai_assistant_service.dart
