+++ flutter_app/lib/features/inventory/models/inventory_models.dart
