+++ flutter_app/lib/features/inventory/screens/add_inventory_item_page.dart
