+++ flutter_app/lib/features/inventory/screens/inventory_page.dart
