+++ flutter_app/lib/features/inventory/services/inventory_service.dart
