+++ flutter_app/lib/features/inventory/widgets/add_item_dialog.dart
