+++ flutter_app/lib/features/inventory/widgets/bottle_card.dart
