+++ flutter_app/lib/features/inventory/widgets/inventory_item_card.dart
