+++ flutter_app/lib/features/inventory/widgets/inventory_shelf.dart
