+++ flutter_app/lib/features/inventory/widgets/lazy_load_section.dart
