+++ flutter_app/lib/features/recipe/recipe.dart
@@ -1,0 +1,2 @@
+// Recipe feature exports
+export 'screens/recipe_screen.dart';
