+++ flutter_app/lib/shared/widgets/connection_line.dart
