+++ flutter_app/test/method_card_comprehensive_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mixologist_flutter/widgets/method_card.dart';
+import 'package:mixologist_flutter/shared/widgets/method_card.dart';
 
 void main() {
   group('MethodCard Comprehensive Tests', () {
