+++ flutter_app/test/method_card_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mixologist_flutter/widgets/method_card.dart';
+import 'package:mixologist_flutter/shared/widgets/method_card.dart';
 
 void main() {
   testWidgets('MethodCard displays title and description', (tester) async {
